Delete XCode project for RakNet
I refuse to support iOS and I especially refuse to deal with the nonsense
that is XCode in its current state. I'll be happy to work with Apple devices
when either I am paid to write code for them or they come to their senses
and do not require paid licenses to get anything working on their platforms.